refactor: Rename special-expressions files now that edits to them are done.
