+++ companion/lib/Controls/Entities/EntitySpecialExpressionManager.ts
