+++ companion/test/Controls/Entities/EntitySpecialExpressionManager.test.ts
@@ -5,7 +5,7 @@ import type { ControlEntityInstance } from '../../../lib/Controls/Entities/Entit
 import {
 	type CreateVariablesAndExpressionParser,
 	EntityPoolSpecialExpressionManager,
-} from '../../../lib/Controls/Entities/EntityIsInvertedManager.js'
+} from '../../../lib/Controls/Entities/EntitySpecialExpressionManager.js'
 import type {
 	UpdateSpecialExpressionValuesFn,
 	NewSpecialExpressionValue,
